requirements.txt
# Flask & Flask-related packages
Click==7.0
Flask==1.0.2
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
Werkzeug==0.14.1

requests
maxminddb-geolite2